pyproject.toml
[tool.poetry]
name = "models"
version = "0.1.0"
description = ""
authors = ["databorodata <nboldyrev90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
django-imagekit = "^5.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
